Start to use fromfile_prefix_chars='@' as argparse.ArgumentParser option.
It is possible to keep a list of arguments in a file that has a file name starting with "@".
The options should then be seperated by newlines.

Update fastPSSM_DOCKER/DOCKER.readme and describe the new command line usage.

The parent git commit message was not quite accurate. The options
 --pfam-dir, --pfamscan-script and --uniprot-db-fasta are also required
when using the software.